updated xmind to 3.5.3

+++ xmind/tools/chocolateyInstall.ps1
@@ -1,18 +1,10 @@
 ﻿$packageName = "xmind"
 $installerType = "EXE"
 $installerArgs = "/SILENT"
-$url = "http://dl2.xmind.net/xmind-downloads/xmind-windows-3.5.1.201411201906.exe"
+$url = "http://dl2.xmind.net/xmind-downloads/xmind-windows-3.5.3.201506180105.exe"
 
-try
-{
-    Install-ChocolateyPackage $packageName $installerType $installerArgs $url
+Install-ChocolateyPackage $packageName $installerType $installerArgs $url
 
+if (Test-Path "$($env:USERPROFILE)\Desktop\XMind 6.lnk") {
     Remove-Item "$($env:USERPROFILE)\Desktop\XMind 6.lnk" -force
-
-    Write-ChocolateySuccess $packageName
-}
-catch
-{
-    Write-ChocolateyFailure $packageName $($_.Exception.Message)
-    throw
 }
